Move getSafeAuthRedirectPath to helpers
