+++ apps/antalmanac/src/backend/lib/helpers.ts
@@ -1,4 +1,7 @@
+import { oidcOAuthEnvSchema } from '$src/backend/env';
 import { TRPCError } from '@trpc/server';
+
+const { GOOGLE_REDIRECT_URI } = oidcOAuthEnvSchema.parse(process.env);
 
 interface FetchAAPIOptions {
     isApiKeyRequired?: boolean;
@@ -129,4 +132,42 @@ export const getCookiesFromHeader = (headers: Headers) => {
             })
     );
     return cookies;
+};
+
+export const getSafeAuthRedirectPath = (redirectUrl: string | null | undefined, requestUrl: string): string => {
+    if (!redirectUrl) {
+        return '/';
+    }
+
+    let requestOrigin: string;
+    try {
+        requestOrigin = new URL(requestUrl).origin;
+    } catch {
+        requestOrigin = new URL(GOOGLE_REDIRECT_URI).origin;
+    }
+
+    const candidates: string[] = [];
+    try {
+        const decodedRedirectUrl = decodeURIComponent(redirectUrl);
+        candidates.push(decodedRedirectUrl);
+    } catch {
+        // Ignore malformed encoding and continue with the raw value.
+    }
+
+    if (!candidates.includes(redirectUrl)) {
+        candidates.push(redirectUrl);
+    }
+
+    for (const candidate of candidates) {
+        try {
+            const parsedRedirectUrl = new URL(candidate, requestOrigin);
+            if (parsedRedirectUrl.origin === requestOrigin) {
+                return `${parsedRedirectUrl.pathname}${parsedRedirectUrl.search}${parsedRedirectUrl.hash}`;
+            }
+        } catch {
+            // Ignore malformed candidate and try next.
+        }
+    }
+
+    return '/';
 };
